pyproject.toml
[project]
name = "automated-curation-platform"
version = "0.1.0"
description = "The ACP was designed to streamline and automate the entire process of deposit package creation, addressing the unique needs of diverse repositories by organizing, filtering, and selecting data accordingly. Leveraging advanced algorithms, machine learning techniques, and predefined rule sets, ACP performs complex curation tasks autonomously, greatly reducing or even eliminating the need for manual input. These automated tasks encompass data gathering, tagging, sorting, and content analysis, which collectively improve efficiency and accuracy."
authors = [
    {name = "Eko Indarto", email = "eko.indarto@dans.knaw.nl"}
]
readme = "README.md"
title = "Automated Curation Platform"
requires-python = ">=3.12"
dependencies = [
    "akmi-utils>=0.1.5.2",
    "asyncio>=3.4.3",
    "boto3>=1.37.5",
    "cryptography>=44.0.2",
    "dynaconf>=3.2.10",
    "emoji>=2.14.1",
    "fastapi-tusd>=0.100.2",
    "jsoncomparison>=1.1.0",
    "jsonpatch>=1.33",
    "psutil>=7.0.0",
    "pydantic>=2.10.6",
    "python-keycloak>=5.3.1",
    "python-multipart>=0.0.20",
    "sickle>=0.7.0",
    "simple-file-checksum>=1.3.1",
    "sqlmodel>=0.0.23",
    "swh-deposit>=1.3.3",
    "sword2>=0.3",
]
